+++ app/src/main/java/com/mlsdev/mlsdevstore/data/local/database/AppDatabase.kt
@@ -4,13 +4,11 @@ import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import com.mlsdev.mlsdevstore.data.local.database.converter.CategoryConverter
-import com.mlsdev.mlsdevstore.data.local.database.dao.AddressDao
-import com.mlsdev.mlsdevstore.data.local.database.dao.CategoriesDao
-import com.mlsdev.mlsdevstore.data.local.database.dao.CheckoutSessionDao
-import com.mlsdev.mlsdevstore.data.local.database.dao.PersonalInfoDao
+import com.mlsdev.mlsdevstore.data.local.database.dao.*
 import com.mlsdev.mlsdevstore.data.model.category.Category
 import com.mlsdev.mlsdevstore.data.model.category.CategoryTree
 import com.mlsdev.mlsdevstore.data.model.category.CategoryTreeNode
+import com.mlsdev.mlsdevstore.data.model.image.CategoryImageEntity
 import com.mlsdev.mlsdevstore.data.model.order.GuestCheckoutSession
 import com.mlsdev.mlsdevstore.data.model.user.Address
 import com.mlsdev.mlsdevstore.data.model.user.PersonalInfo
@@ -21,7 +19,8 @@ import com.mlsdev.mlsdevstore.data.model.user.PersonalInfo
     CategoryTree::class,
     Address::class,
     PersonalInfo::class,
-    GuestCheckoutSession::class
+    GuestCheckoutSession::class,
+    CategoryImageEntity::class
 ], version = 1, exportSchema = false)
 @TypeConverters(CategoryConverter::class)
 abstract class AppDatabase : RoomDatabase() {
@@ -29,4 +28,5 @@ abstract class AppDatabase : RoomDatabase() {
     abstract fun addressDao(): AddressDao
     abstract fun personalInfoDao(): PersonalInfoDao
     abstract fun checkoutSessionDao(): CheckoutSessionDao
+    abstract fun categoryImagesDao(): CategoryImagesDao
 }