+++ app/src/main/java/com/mlsdev/mlsdevstore/data/local/database/Table.kt
@@ -8,6 +8,7 @@ annotation class Table {
         const val ADDRESSES = "addresses"
         const val PERSONAL_INFO = "personal_info"
         const val GUEST_CHECKOUT_SESSION = "guest_checkout_session"
+        const val CATEGORY_IMAGES = "category_images"
     }
 
 }