+++ app/src/main/java/com/mlsdev/mlsdevstore/injections/module/ApiModule.kt
@@ -4,10 +4,7 @@ import com.google.gson.Gson
 import com.mlsdev.mlsdevstore.BuildConfig
 import com.mlsdev.mlsdevstore.data.local.SharedPreferencesManager
 import com.mlsdev.mlsdevstore.data.remote.AuthInterceptor
-import com.mlsdev.mlsdevstore.data.remote.service.AuthenticationService
-import com.mlsdev.mlsdevstore.data.remote.service.BrowseService
-import com.mlsdev.mlsdevstore.data.remote.service.OrderService
-import com.mlsdev.mlsdevstore.data.remote.service.TaxonomyService
+import com.mlsdev.mlsdevstore.data.remote.service.*
 import com.mlsdev.mlsdevstore.injections.Named
 import dagger.Module
 import dagger.Provides
@@ -25,6 +22,7 @@ class ApiModule {
     annotation class Name {
         companion object {
             const val WITHOUT_AUTHORIZATION_INTERCEPTOR = "with_authorization_interceptor"
+            const val SEARCH_IMAGE_RETROFIT = "search_image_retrofit"
         }
     }
 
@@ -46,6 +44,18 @@ class ApiModule {
     private fun createRetrofit(client: OkHttpClient, gson: Gson): Retrofit {
         return Retrofit.Builder()
                 .baseUrl(BuildConfig.BASE_URL)
+                .client(client)
+                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
+                .addConverterFactory(GsonConverterFactory.create(gson))
+                .build()
+    }
+
+    @Provides
+    @Singleton
+    @Named(Name.SEARCH_IMAGE_RETROFIT)
+    fun provideSearchImageApiRetrofit(@Named(Name.WITHOUT_AUTHORIZATION_INTERCEPTOR) client: OkHttpClient, gson: Gson): Retrofit {
+        return Retrofit.Builder()
+                .baseUrl("https://pixabay.com/api/")
                 .client(client)
                 .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
                 .addConverterFactory(GsonConverterFactory.create(gson))
@@ -106,5 +116,9 @@ class ApiModule {
     fun provideAuthInterceptor(sharedPreferencesManager: SharedPreferencesManager): AuthInterceptor {
         return AuthInterceptor(sharedPreferencesManager)
     }
+
+    @Provides
+    fun provideSearchImageService(@Named(Name.SEARCH_IMAGE_RETROFIT) retrofit: Retrofit): SearchImageService =
+            retrofit.create(SearchImageService::class.java)
 
 }